requirements.txt
altair
pandas
streamlit
opencv-python==4.11.0.86
numpy==1.26.4
matplotlib==3.7.2
scikit-image==0.25.2
pandas==2.2.3
tensorflow==2.18.0
pillow==11.1.0
seaborn==0.12.2
plotly==5.24.1